assetto-corsa-rl/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "assetto-corsa-rl"
version = "0.1.7"
description = "Reinforcement Learning toolkit for Assetto Corsa and Car Racing environments"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Assetto Corsa RL Contributors" }]
keywords = [
    "reinforcement-learning",
    "assetto-corsa",
    "racing",
    "SAC",
    "VAE",
    "deep-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Games/Entertainment :: Simulation",
]

dependencies = [
    "torch>=2.0.0",
    "torchrl>=0.2.0",
    "tensordict>=0.2.0",
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "matplotlib>=3.7.0",
    "click>=8.1.0",
    "PyYAML>=6.0",
    "lightning>=2.0.0",
    "wandb>=0.15.0",
    "vgamepad>=0.1.0",
    "mss>=10.0.0",
    "pywin32>=223",
    "lpips>=0.1.4",
    "torchinfo>=1.8.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://assettocorsarl.github.io/AssettoCorsaRL-DOCS/docs/"
Repository = "https://github.com/AssettoCorsaRL/AssettoCorsaRL"
Documentation = "https://assettocorsarl.github.io/AssettoCorsaRL-DOCS/docs/"
"Bug Tracker" = "https://github.com/AssettoCorsaRL/AssettoCorsaRL/issues"

[project.scripts]
acrl = "assetto_corsa_rl.cli:main"

[tool.setuptools]
packages = { find = { where = ["src"] } }
include-package-data = true

[tool.setuptools.package-data]
"assetto_corsa_rl" = ["configs/**/*.yaml"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=assetto_corsa_rl --cov-report=term-missing"
